wip: wsl2 diagnostics script
