+++ .editorconfig
@@ -1,0 +1,11 @@
+root = true
+
+[*]
+end_of_line = lf
+max_line_length = 120
+indent_size = 2
+tab_width = 2
+indent_style = space
+charset = utf-8
+insert_final_newline = true
+trim_trailing_whitespace = true
